Moves ui5 releated code to own folder
