+++ packages/fiori-generator-shared/src/index.ts
@@ -5,8 +5,8 @@ export * from './system-utils';
 export * from './telemetry';
 export * from './logging/logWrapper';
 export { getPackageScripts } from './npm-package-scripts/getPackageScripts';
-export { getBootstrapResourceUrls } from './helpers';
+export { getBootstrapResourceUrls } from './ui5/ui5';
 export { getDefaultTargetFolder, isExtensionInstalled } from './vscode-helpers/vscode-helpers';
 export { generateReadMe } from './read-me/read-me';
 export { getHostEnvironment } from './environment';
-export { YeomanEnvironment, VSCodeInstance, hostEnvironment } from './types';
+export { YeomanEnvironment, VSCodeInstance, hostEnvironment, HostEnvironmentId } from './types';
