--- packages/fiori-generator-shared/src/utility/readme.md
@@ -1,1 +1,0 @@
-Add generally applicable functions here to promote reuse rather.
